feat: Moves compress from parquet workspace to arrow-buffer
