+++ parquet/src/arrow/record_reader/definition_levels.rs
@@ -15,11 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use arrow_array::builder::BooleanBufferBuilder;
-use arrow_buffer::Buffer;
-use arrow_buffer::bit_chunk_iterator::UnalignedBitChunk;
-use bytes::Bytes;
-
 use crate::arrow::buffer::bit_util::count_set_bits;
 use crate::basic::Encoding;
 use crate::column::reader::decoder::{
@@ -27,6 +22,11 @@ use crate::column::reader::decoder::{
 };
 use crate::errors::{ParquetError, Result};
 use crate::schema::types::ColumnDescPtr;
+use arrow_array::builder::BooleanBufferBuilder;
+use arrow_buffer::Buffer;
+use arrow_buffer::bit_chunk_iterator::UnalignedBitChunk;
+use arrow_buffer::bit_util::compress;
+use bytes::Bytes;
 
 enum BufferInner {
     /// Compute levels and null mask
@@ -213,8 +213,6 @@ pub(crate) fn build_filtered_validity_bitmap(
 
     item_count
 }
-
-use crate::util::bit_util::compress;
 
 enum MaybePacked {
     Packed(PackedDecoder),
